smaller config fixes

+++ package/upload.sh
@@ -1,3 +1,8 @@
+source .env
+if [ -z ${PYPI_TOKEN+x} ]; then
+	echo 'Cannot upload without a pypi token. Please set one in a .env file next to this script with the name of `PYPI_TOKEN`.'
+	exit 1
+fi
 echo "🧽 Pruging old package files... 🧽"
 rm -rf dist
 mkdir dist
@@ -6,6 +11,6 @@ black . > /dev/null
 echo "🛠  Building... 🛠"
 poetry build > /dev/null
 echo -n "☁️  Uploading... ☁️"
-poetry publish -u __token__ -p $TOKEN
-#python3 -m twine upload -r pypi dist/*
+poetry publish -u __token__ -p $PYPI_TOKEN
 echo "✨ All done! ✨"
+unset PYPI_TOKEN
